valinet-unserver: Support for packages apps
Some functionality of packaged apps (e.g. Modern UI apps) is hosted in
a separate `RuntimeBroker.exe` process. This commit adds support for
"unservering" such apps as well: mod injects all `RuntimeBroker.exe`
processes, determines the package name of the current process and
checks if the beginning of it matches any of the strings listed in the
`runtimeBrokerHooksPackageName` list at the top of the module.
If the injected process doesn't belong to a package we want injected,
`Wh_Init` simply returns FALSE, thus unloading the module. Otherwise,
that process gets patched, i.e. "unservered".

Make sure to add `RuntimeBroker.exe` in "Process inclusion list" in
Windhawk's advanced settings. Also, strings in the
`runtimeBrokerHooksPackageName` list have to be all lower case.

This has been made so that the Start menu could finally display sleep
options (like "Sleep", "Hibernate" etc) in the power menu. For some
reason only God knows, there is a hardcoded check in `shutdownux.dll`
(the DLL which hosts the power menu logic) for OS_ANYSERVER, and if it
sees it's running on Windows Server, it simply hides those options
despite the actual availability of those options on the host - mind
blowing. That' why the desktop Alt+F4 Shutdown Dialog coded properly
30 years ago can display the options, but the "modern" Start menu
cannot. Relevant code is in: `CShutdownChoices::GetDefaultUIChoiceMask`:

```
  v9 = SHWindowsPolicy(POLID_NoClose, v8);
  if ( !v6 && !v9 && (IsOS(29u) || !v7 || v4 || !CShutdownChoices::_IsBitlockerEnabledForOSPartition(this)) )
  {
    v10 = IsOS(29u);
    v11 = *a2;
    if ( v10 )
      v12 = v11 | 0x20006;
    else
      v12 = v11 | 0x8020056;
    *a2 = v12;
  }
```

If `IsOS(29)` (aka `IsOS(OS_ANYSERVER)`) is patched to return 0 instead
of 1, "Hibernate" (which is enabled on my server) displays just fine.

Mind blowing.