valinet-chrome-save-to-downloads-directly: Initial version
This module bypasses the "Save As" dialog and directly saves the file
to whatever folder the dialog starts in (usually Downloads) in Chrome.
This is a hack so that file downloads start right away when having Safe
Browsing completely disabled (which normally produces some idiotic
confirmation popup before downloading any file). This module has to be
used alongside the "Ask where to save each file before downloading"
setting and basically restores the normal download experience.